Huge QoL improvements to content update page

+++ frontend/src/utils/numbers.ts
@@ -1,20 +1,12 @@
 import * as _ from 'lodash-es';
 
 export function sign(num: number | string): string {
-  num = _.isString(num) ? parseInt(num) : num;
+  num = _.isString(num) ? parseFloat(num) : num;
   return num < 0 ? `${num}` : `+${num}`;
 }
 
 export function rankNumber(num: number, zeroName = '0'): string {
-  return num === 0
-    ? zeroName
-    : num === 1
-    ? '1st'
-    : num === 2
-    ? '2nd'
-    : num === 3
-    ? '3rd'
-    : `${num}th`;
+  return num === 0 ? zeroName : num === 1 ? '1st' : num === 2 ? '2nd' : num === 3 ? '3rd' : `${num}th`;
 }
 
 export function hashData(data: Record<string, any>): number {
